requirements.txt
watchdog==3.0.0
loguru==0.7.0
PyYAML==6.0